Add valid

+++ src/service/commentService.js
@@ -14,7 +14,8 @@ var connection
 
 const createCommentService = async(comment) => {
   const id = uuidv4();
-
+  if(comment.content.length == 0 || comment.postId.length == 0 || comment.userId.length == 0)
+    return 0
   const [results, fields] = await connection.query(
       `insert into Comment(id, content, postId, userId)
        values (?, ?, ?, ?)`,
@@ -25,7 +26,8 @@ const createCommentService = async(comment) => {
 }
 
 const getSingleCommentService = async(id) => {
- 
+  if(id.id.length == 0 )
+    return null
   const [results, fields] = await connection.query(
       `select * from Comment  
       where id = ?`,
@@ -44,7 +46,8 @@ const getAllCommentService = async() => {
 }
 
 const deleteCommentService = async(id) => {
- 
+  if(id.id.length == 0 )
+    return 0
   const [results, fields] = await connection.query(
       `delete from Comment where id = ?`,
       [id.id]  
@@ -53,7 +56,8 @@ const deleteCommentService = async(id) => {
 }
 
 const updateCommentService = async(comment) => {
- 
+  if(comment.content.length == 0 || comment.postId.length == 0 || comment.userId.length == 0 || comment.id.length == 0)
+    return 0
   const [results, fields] = await connection.query(
       `update Post
        set  content = ?, postId = ?, userId = ?
