+++ src/service/userService.js
@@ -1,3 +1,4 @@
+const bcrypt = require('bcrypt');
 const { v4: uuidv4 } = require('uuid');
 const connectionPromise = require('../config/database')
 var connection
@@ -13,7 +14,8 @@ var connection
 })()
 
 const getSingleUserService = async(id) => {
- 
+  if(id.id.length == 0 )
+   return null
   const [results, fields] = await connection.query(
       `select * from User 
       where id = ?`,
@@ -32,7 +34,8 @@ const getAllUserService = async() => {
 }
 
 const deleteUserService = async(id) => {
- 
+  if(id.id.length == 0 )
+    return null
   const [results, fields] = await connection.query(
       `delete from User where id = ?`,
       [id.id]  
@@ -41,12 +44,15 @@ const deleteUserService = async(id) => {
 }
 
 const updateUserService = async(user) => {
- 
+  if(user.name.length == 0 ||  user.email.length == 0 ||  user.password, user.id)
+    return 0
+  const saltRounds = 10;
+  const password = await bcrypt.hash(user.password, saltRounds);
   const [results, fields] = await connection.query(
       `update User
        set name = ?,  email = ?, password = ?
        where id = ?`,
-       [user.name, user.email, user.password, user.id]
+       [user.name, user.email, password, user.id]
          
   )
   return results['affectedRows']
